More progress, new sounds for dialing sequence, dhd doesn't lock until dial
