+++ code/sbox_stargate/entities/stargate_milkyway/StargateRing.cs
@@ -5,190 +5,186 @@ using System.Text;
 using System.Threading.Tasks;
 using Sandbox;
 
-public partial class StargateRing : AnimEntity
+public partial class StargateRing : PlatformEntity
 {
 	// ring variables
 
-	public Stargate Gate;
-	public float RingAngle { get; private set; } = 0.0f;
-	public float TargetRingAngle { get; private set; } = 0.0f;
-	public float RingRotSpeed { get; private set; } = 0.0f;
-	private float RingRotMinSpeed = 0.5f;
-	private float RingRotMaxSpeed = 10f;
-	public int RingRotDir { get; private set; } = 1;
-	public bool RingShouldRotate { get; private set; } = false;
-	private bool RingShouldAcc = false;
-	private bool RingShouldDecc = false;
-	public bool RingFreeSpin { get; private set; } = false;
+	[Net]
+	public Stargate Gate { get; set; } = null;
+
 	public string RingSymbols { get; private set; } = "?0JKNTR3MBZX*H69IGPL#@QFS1E4AU85OCW72YVD";
 
-	private Sound RollSound;
+	[Net]
+	public float RingAngle { get; private set; } = 0.0f;
+	[Net]
+	public char GateDialingSymbol { get; private set; } = '!';
+	public float TargetRingAngle { get; private set; } = 0.0f;
+
+	private float RingCurSpeed = 0f;
+	private float RingMaxSpeed = 50f;
+	private float RingAccelStep = 1f;
+	private float RingDeccelStep = 0.75f;
+
+	private int RingDirection = 1;
+	private bool ShouldAcc = false;
+	private bool ShouldDecc = false;
+
+	private bool ShouldStopAtAngle = false;
+	private float CurStopAtAngle = 0f;
+
+	private float StartedAccelAngle = 0f;
+	private float StoppedAccelAngle = 0f;
+
+	private string StartSoundName = "gate_roll_long";
+	private string LoopSoundName = "gate_sg1_ring_loop";
+	private string StopSoundName = "gate_sg1_ring_stop";
+
+	private Sound? StartSoundInstance;
+	private Sound? LoopSoundInstance;
+	private Sound? StopSoundInstance;
 
 	public override void Spawn()
 	{
-		base.Spawn();
+		Transmit = TransmitType.Always;
 
 		SetModel( "models/gmod4phun/stargate/gate_sg1/ring_sg1.vmdl" );
 
-		Transmit = TransmitType.Always;
+		SpawnSettings = Flags.LoopMovement;
+		MoveDirType = PlatformMoveType.RotatingContinious;
+		MoveDirIsLocal = true;
+		MoveDir = Rotation.Up.EulerAngles;
+		MoveDistance = 360;
+
+		//StartMoveSound = "gate_roll_long";
+		//StartMoveSound = "gate_sg1_ring_start";
+		//MovingSound = "gate_sg1_ring_loop";
+		//StopMoveSound = "gate_sg1_ring_stop";
+
+		base.Spawn();
+
+		EnableAllCollisions = false;
 	}
 
-	public void FlipRingRotDir()
+	protected override void OnDestroy()
 	{
-		RingRotDir = -RingRotDir;
+		if ( StartSoundInstance.HasValue ) StartSoundInstance.Value.Stop();
+		if ( LoopSoundInstance.HasValue ) LoopSoundInstance.Value.Stop();
+		if ( StopSoundInstance.HasValue ) StopSoundInstance.Value.Stop();
+
+		base.OnDestroy();
 	}
 
-	public void PlayRollSound()
-	{
-		StopRollSound();
-		RollSound = Sound.FromEntity( "gate_roll_long", this );
-	}
-
-	public void StopRollSound()
-	{
-		RollSound.Stop();
-	}
-
-	public void StartRingRotation() // start ring free spin rotation
-	{
-		FlipRingRotDir(); // swap the direction on each rotation call
-
-		RingShouldRotate = true;
-		RingFreeSpin = true;
-		RingShouldAcc = true;
-		RingShouldDecc = false;
-
-		PlayRollSound();
-	}
-
-	public void StopRingRotation() // stop ring free spin rotation
-	{
-		RingShouldRotate = true;
-		RingFreeSpin = true;
-		RingShouldAcc = false;
-		RingShouldDecc = true;
-
-		StopRollSound();
-	}
-
-	public void RotateRingTo( float targetAng ) // gradually rotates the ring to a specific angle
-	{
-		FlipRingRotDir(); // swap the direction on each rotation call
-
-		TargetRingAngle = targetAng;
-		RingShouldRotate = true;
-		RingFreeSpin = false;
-		RingShouldAcc = true;
-		RingShouldDecc = false;
-
-		PlayRollSound();
-	}
-
-	public void StopRingRotationInstant() // instantly stops the ring rotation
-	{
-		RingRotSpeed = 0;
-		RingShouldRotate = false;
-		RingFreeSpin = false;
-		RingShouldAcc = false;
-		RingShouldDecc = false;
-
-		StopRollSound();
-	}
-
-	public void ResetRingRotation() // stops rotation and resets the ring angle back to 0
-	{
-		RingAngle = 0;
-		TargetRingAngle = 0;
-		RingRotSpeed = 0;
-		RingShouldRotate = false;
-		RingShouldAcc = false;
-		RingShouldDecc = false;
-	}
-
-	private void RingRotationLogic() // think logic for the ring rotation
-	{
-		if ( RingShouldRotate )
-		{
-			var angDiff = Math.Abs( RingAngle - TargetRingAngle );
-
-			var angDeltaCalc = (1 / ((angDiff * 0.1f) + 1)) - 0.05f;
-
-			if ( angDiff < 40f && RingShouldAcc && !RingFreeSpin )
-			{
-				RingShouldAcc = false;
-				RingShouldDecc = true;
-			}
-
-			if ( RingShouldAcc && !RingShouldDecc && (RingRotSpeed != RingRotMaxSpeed) )
-			{
-				RingRotSpeed = RingRotSpeed.LerpTo( RingRotMaxSpeed, Time.Delta * 1.5f );
-				if ( (RingRotMaxSpeed - RingRotSpeed) < 0.05 ) RingRotSpeed = RingRotMaxSpeed;
-				//Log.Info($"Accelerating, curSpeed = {RingRotSpeed}");
-			}
-			else if ( RingShouldDecc && !RingShouldAcc && (RingRotSpeed != RingRotMinSpeed) )
-			{
-				if (!RingFreeSpin)
-				{
-					RingRotSpeed = RingRotSpeed.LerpTo( RingRotMinSpeed, Time.Delta * angDeltaCalc * 2.5f);
-					if ( (RingRotSpeed - RingRotMinSpeed) < 0.05 ) RingRotSpeed = RingRotMinSpeed;
-				}
-				else
-				{
-					RingRotSpeed = RingRotSpeed.LerpTo( 0, Time.Delta * 2 );
-					if ( Math.Abs( RingRotSpeed - 0 ) < 0.05 ) RingRotSpeed = 0;
-				}
-				//Log.Info( $"Deccelerating, curSpeed = {RingRotSpeed}" );
-			}
-
-			if (!RingFreeSpin)
-			{
-				if ( TargetRingAngle > RingAngle ) RingAngle += 0.1f * RingRotSpeed;
-				else if ( TargetRingAngle < RingAngle ) RingAngle -= 0.1f * RingRotSpeed;
-			}
-			else
-			{
-				RingAngle += 0.1f * RingRotSpeed * -RingRotDir;
-				if (RingAngle > 360f || RingAngle < 0f) RingAngle = RingAngle.UnsignedMod( 360f );
-			}
-
-			if ( RingShouldDecc && angDiff < 0.1f )
-			{
-				if ( angDiff != 0 && !RingFreeSpin)
-				{
-					RingAngle = TargetRingAngle;
-					StopRingRotationInstant();
-				}
-			}
-
-			if ( RingFreeSpin && RingRotSpeed == 0 )
-			{
-				RingShouldRotate = false;
-				RingAngle = RingAngle.UnsignedMod( 360 );
-			}
-
-			LocalRotation = LocalRotation.Angles().WithRoll( RingAngle ).ToRotation();
-		}
-	}
-
-	public float GetSymbolPosition(char sym) // gets the symbols position on the ring
+	// symbol pos/ang
+	public float GetSymbolPosition( char sym ) // gets the symbols position on the ring
 	{
 		sym = sym.ToString().ToUpper()[0];
 		return RingSymbols.Contains( sym ) ? RingSymbols.IndexOf( sym ) : -1;
 	}
 
-	public float GetDesiredRingAngleForSymbol(char sym, int angOffset = 0)
+	public float GetSymbolAngle( char sym ) // gets the symbols angle on the ring
+	{
+		sym = sym.ToString().ToUpper()[0];
+		return GetSymbolPosition( sym ) * 9;
+	}
+
+	// sounds
+	public void StopStartSound()
+	{
+		if ( StartSoundInstance.HasValue ) StartSoundInstance.Value.Stop();
+	}
+
+	public void PlayStartSound()
+	{
+		StopStartSound();
+		StartSoundInstance = PlaySound( StartSoundName );
+	}
+
+	public void StopStopSound()
+	{
+		if ( StopSoundInstance.HasValue ) StopSoundInstance.Value.Stop();
+	}
+
+	public void PlayStopSound()
+	{
+		StopStopSound();
+		StopSoundInstance = PlaySound( StopSoundName );
+	}
+
+	public async void StopLoopSound( float delay = 0 )
+	{
+		if (delay > 0)
+		{
+			await Task.DelaySeconds( delay );
+			if ( !this.IsValid() ) return;
+		}
+
+		if ( LoopSoundInstance.HasValue ) LoopSoundInstance.Value.Stop();
+	}
+
+	public async void PlayLoopSound( float delay = 0 )
+	{
+		if ( delay > 0 )
+		{
+			await Task.DelaySeconds( delay );
+			if ( !this.IsValid() ) return;
+		}
+
+		StopLoopSound();
+		LoopSoundInstance = PlaySound( LoopSoundName );
+	}
+
+	// spinup/spindown - starts or stops rotating the ring
+	public void SpinUp()
+	{
+		ShouldDecc = false;
+		ShouldAcc = true;
+	}
+
+	public void SpinDown()
+	{
+		ShouldAcc = false;
+		ShouldDecc = true;
+	}
+
+	public void OnStart()
+	{
+		PlayStartSound();
+	}
+
+	public void OnStop()
+	{
+		if (Gate.IsValid() && Gate.Dialing) { PlayStopSound(); }
+		
+		StopStartSound();
+	}
+
+	// rotate to angle/symbol
+	public void RotateRingTo( float targetAng ) // starts rotating the ring and stops (hopefully) at the specified angle
+	{
+		TargetRingAngle = targetAng;
+		ShouldStopAtAngle = true;
+		SpinUp();
+	}
+
+	public void RotateRingToSymbol( char sym, int angOffset = 0 )
+	{
+		if ( RingSymbols.Contains( sym ) ) RotateRingTo( GetDesiredRingAngleForSymbol( sym, angOffset ) );
+	}
+
+	// helper calcs
+	public float GetDesiredRingAngleForSymbol( char sym, int angOffset = 0 )
 	{
 		// get the symbol's position on the ring
-		var symPos = GetSymbolPosition(sym);
+		var symPos = GetSymbolPosition( sym );
 
 		// if we input an invalid symbol, return current ring angles
-		if (symPos == -1) return RingAngle;
+		if ( symPos == -1 ) return RingAngle;
 
 		// if its a valid symbol, lets calc the required angle
 		var symAng = symPos * 9; // there are 40 symbols, each 9 degrees apart
 
 		// clockwise and counterclockwise symbol angles relative to 0 (the top chevron)
-		var D_CW = symAng - RingAngle - angOffset; // offset, if we want it to be relative to another chevron (for movie stargate dialing)
+		var D_CW = -symAng - RingAngle - angOffset; // offset, if we want it to be relative to another chevron (for movie stargate dialing)
 		var D_CCW = 360 - D_CW;
 
 		D_CW = D_CW.UnsignedMod( 360 );
@@ -196,40 +192,32 @@ public partial class StargateRing : AnimEntity
 
 		// angle differences are setup, choose based on the direction of ring rotation
 		// if the required angle to too small, spin it around once
-		var angToRotate = (RingRotDir == -1) ? D_CCW : D_CW;
-		if (angToRotate < 170f) angToRotate += 360f;
+		var angToRotate = (RingDirection == -1) ? D_CCW : D_CW;
+		if ( angToRotate < 170f ) angToRotate += 360f;
 
 		// set the final angle to the current angle + the angle needed to rotate, also considering ring direction
-		var finalAng = RingAngle + (angToRotate * RingRotDir);
+		var finalAng = RingAngle + (angToRotate * RingDirection);
 
 		//Log.Info($"Sym = {sym}, RingAng = {RingAngle}, SymPos = {symPos}, D_CCW = {D_CCW}, D_CW = {D_CW}, finalAng = {finalAng}" );
 
 		return finalAng;
 	}
 
-	public void RotateRingToSymbol(char sym, int angOffset = 0)
+	public async Task<bool> RotateRingToSymbolAsync( char sym, int angOffset = 0 )
 	{
-		if ( RingSymbols.Contains( sym ) ) RotateRingTo( GetDesiredRingAngleForSymbol( sym, angOffset ) );
-	}
+		RotateRingToSymbol( sym, angOffset );
+		GateDialingSymbol = sym;
 
-	public async Task<bool> RotateRingToSymbolAsync(char sym, int angOffset = 0)
-	{
-		RotateRingToSymbol( sym, angOffset ); // rotate ring to the desired position
+		await Task.DelaySeconds( Global.TickInterval ); // wait, otherwise it hasnt started moving yet and can cause issues
 
-		var rotStartTime = Time.Now;
-
-		while ( RingAngle != TargetRingAngle ) // wait until its rotated
+		while (IsMoving)
 		{
-			await Task.DelaySeconds( 0.1f ); // we need to delay this otherwise the game hangs
+			await Task.DelaySeconds( Global.TickInterval ); // wait here, too, otherwise game hangs :)
 
-			if ( Time.Now > rotStartTime + 30 ) return false;
-
-			if ( !this.IsValid() || !Gate.IsValid() ) return false;
-
-			if (Gate.Dialing && Gate.ShouldStopDialing)
+			if ( Gate.ShouldStopDialing )
 			{
-				Gate.ShouldStopDialing = false;
-				StopRingRotation();
+				SpinDown();
+				Gate.CurGateState = Stargate.GateState.IDLE;
 				return false;
 			}
 		}
@@ -237,38 +225,96 @@ public partial class StargateRing : AnimEntity
 		return true;
 	}
 
-	[Event( "server.tick" )]
+	[Event.Tick.Server]
 	public void Think()
 	{
-		RingRotationLogic();
+		if ( !Gate.IsValid() ) return;
+
+		if ( IsMoving && Gate.ShouldStopDialing )
+		{
+			SpinDown();
+			Gate.CurGateState = Stargate.GateState.IDLE;
+		}
+
+		if ( ShouldAcc )
+		{
+			if ( !IsMoving )
+			{
+				StartedAccelAngle = RingAngle;
+				StartMoving();
+				OnStart();
+			}
+
+			if ( RingCurSpeed < RingMaxSpeed )
+			{
+				RingCurSpeed += RingAccelStep;
+			}
+			else
+			{
+				RingCurSpeed = RingMaxSpeed;
+				ShouldAcc = false;
+				StoppedAccelAngle = MathF.Abs( RingAngle - StartedAccelAngle );
+				CurStopAtAngle = TargetRingAngle - (StoppedAccelAngle * RingDirection * (RingAccelStep / RingDeccelStep));
+			}
+		}
+		else if ( ShouldDecc )
+		{
+			if ( RingCurSpeed > 0 )
+			{
+				RingCurSpeed -= RingDeccelStep;
+			}
+			else
+			{
+				RingCurSpeed = 0;
+				ShouldDecc = false;
+				StopMoving();
+				OnStop();
+
+				ReverseMoving();
+				CurrentRotation %= 360f;
+			}
+		}
+
+		SetSpeed( RingCurSpeed );
+
+		if (ShouldStopAtAngle && IsMoving)
+		{
+			if ( !ShouldAcc && !ShouldDecc )
+			{
+				var angDiff = MathF.Abs( CurrentRotation - CurStopAtAngle );
+				if (angDiff < 1f)
+				{
+					SpinDown();
+					ShouldStopAtAngle = false;
+				}
+			}
+		}
+
+		RingAngle = CurrentRotation;
+		RingDirection = IsMovingForwards ? -1 : 1;
 	}
 
-	protected override void OnDestroy()
-	{
-		base.OnDestroy();
-
-		StopRollSound();
-	}
-
-
+	// DEBUG
 	public void DrawSymbols()
 	{
 		var deg = 360f / RingSymbols.Length;
 		var i = 0;
 		var ang = Rotation.Angles();
-		foreach (char sym in RingSymbols)
+		foreach ( char sym in RingSymbols )
 		{
 			var rotAng = ang.WithRoll( ang.roll - (i * deg) );
 			var newRot = rotAng.ToRotation();
 			var pos = Position + newRot.Forward * 4 + newRot.Up * 117.5f;
-			DebugOverlay.Text( pos, sym.ToString(), Color.Yellow );
+			DebugOverlay.Text( pos, sym.ToString(), sym == GateDialingSymbol ? Color.Green : Color.Yellow );
 			i++;
 		}
+
+		if (Gate.IsValid())	DebugOverlay.Text( Position, Gate.ShouldStopDialing.ToString(), Color.White );
 	}
 
 	[Event.Frame]
 	public void RingSymbolsDebug()
 	{
-		//if ( Local.Pawn.IsValid() && Local.Pawn.Position.Distance( Position ) < 800 ) DrawSymbols();
+		if ( Local.Pawn.IsValid() && Local.Pawn.Position.Distance( Position ) < 800 ) DrawSymbols();
 	}
 }
