+++ sounds/gmod4phun/stargate/dhd/sg1/dhd_sg1_press.sound
@@ -13,7 +13,7 @@
 			"sounds/gmod4phun/stargate/dhd/sg1/press_7.vsnd",
 		]
 		UI = false
-		Volume = 1.0
+		Volume = 0.5
 		VolumeRandom = 0.0
 		Pitch = 1.0
 		PitchRandom = 0.0
