+++ sounds/gmod4phun/stargate/dial_fail_sg1.sound
@@ -6,5 +6,11 @@
 		[
 			"sounds/gmod4phun/stargate/dial_fail_sg1.vsnd",
 		]
+		UI = false
+		Volume = 0.5
+		VolumeRandom = 0.0
+		Pitch = 1.0
+		PitchRandom = 0.0
+		DistanceMax = 1000.0
 	}
 }
