+++ sounds/sbox_stargate/stargate/sg1/ring/gate_sg1_ring_stop.sound
@@ -6,5 +6,11 @@
 		[
 			"sounds/sbox_stargate/stargate/sg1/ring/spinstop.vsnd",
 		]
+		UI = false
+		Volume = 1.0
+		VolumeRandom = 0.0
+		Pitch = 1.0
+		PitchRandom = 0.0
+		DistanceMax = 1000.0
 	}
 }
